wp updated

+++ wordpress/wp-content/plugins/wordfence/lib/wfConfig.php
@@ -34,6 +34,7 @@ class wfConfig {
 				"scansEnabled_plugins" => false,
 				"scansEnabled_malware" => false,
 				"scansEnabled_fileContents" => false,
+				"scansEnabled_database" => false,
 				"scansEnabled_posts" => false,
 				"scansEnabled_comments" => false,
 				"scansEnabled_passwds" => false,
@@ -64,8 +65,12 @@ class wfConfig {
 				"startScansRemotely" => false,
 				"disableConfigCaching" => false,
 				"addCacheComment" => false,
+				"disableCodeExecutionUploads" => false,
 				"allowHTTPSCaching" => false,
-				"debugOn" => false
+				"debugOn" => false,
+				'email_summary_enabled' => true,
+				'email_summary_dashboard_widget_enabled' => true,
+				'ssl_verify' => true,
 			),
 			"otherParams" => array(
 				'securityLevel' => '0',
@@ -88,7 +93,10 @@ class wfConfig {
 				'max404Humans_action' => "throttle",
 				'maxScanHits' => "DISABLED",
 				'maxScanHits_action' => "throttle",
-				'blockedTime' => "300"
+				'blockedTime' => "300",
+				'email_summary_interval' => 'biweekly',
+				'email_summary_excluded_directories' => 'wp-content/cache,wp-content/wfcache,wp-content/plugins/wordfence/tmp',
+				'allowed404s' => "/favicon.ico\n/apple-touch-icon*.png\n/*@2x.png",
 			)
 		),
 		array( //level 1
@@ -116,88 +124,7 @@ class wfConfig {
 				"scansEnabled_plugins" => false,
 				"scansEnabled_malware" => true,
 				"scansEnabled_fileContents" => true,
-				"scansEnabled_posts" => true,
-				"scansEnabled_comments" => true,
-				"scansEnabled_passwds" => true,
-				"scansEnabled_diskSpace" => true,
-				"scansEnabled_options" => true,
-				"scansEnabled_dns" => true,
-				"scansEnabled_scanImages" => false,
-				"scansEnabled_highSense" => false,
-				"scansEnabled_oldVersions" => true,
-				"firewallEnabled" => true,
-				"blockFakeBots" => false,
-				"autoBlockScanners" => true,
-				"loginSecurityEnabled" => true,
-				"loginSec_lockInvalidUsers" => false,
-				"loginSec_maskLoginErrors" => true,
-				"loginSec_blockAdminReg" => true,
-				"loginSec_disableAuthorScan" => true,
-				"other_hideWPVersion" => true,
-				"other_noAnonMemberComments" => true,
-				"other_blockBadPOST" => false,
-				"other_scanComments" => true,
-				"other_pwStrengthOnUpdate" => true,
-				"other_WFNet" => true,
-				"other_scanOutside" => false,
-				"deleteTablesOnDeact" => false,
-				"autoUpdate" => false,
-				"disableCookies" => false,
-				"startScansRemotely" => false,
-				"disableConfigCaching" => false,
-				"addCacheComment" => false,
-				"allowHTTPSCaching" => false,
-				"debugOn" => false
-			),
-			"otherParams" => array(
-				'securityLevel' => '1',
-				"alertEmails" => "", "liveTraf_ignoreUsers" => "", "liveTraf_ignoreIPs" => "", "liveTraf_ignoreUA" => "",  "apiKey" => "", "maxMem" => '256', 'scan_exclude' => '', 'whitelisted' => '', 'bannedURLs' => '', 'maxExecutionTime' => '', 'howGetIPs' => '', 'actUpdateInterval' => '', 'alert_maxHourly' => 0, 'loginSec_userBlacklist' => '',
-				"neverBlockBG" => "neverBlockVerified",
-				"loginSec_countFailMins" => "5",
-				"loginSec_lockoutMins" => "5",
-				'loginSec_strongPasswds' => 'pubs',
-				'loginSec_maxFailures' => "50",
-				'loginSec_maxForgotPasswd' => "50",
-				'maxGlobalRequests' => "DISABLED",
-				'maxGlobalRequests_action' => "throttle",
-				'maxRequestsCrawlers' => "DISABLED",
-				'maxRequestsCrawlers_action' => "throttle",
-				'maxRequestsHumans' => "DISABLED",
-				'maxRequestsHumans_action' => "throttle",
-				'max404Crawlers' => "DISABLED",
-				'max404Crawlers_action' => "throttle",
-				'max404Humans' => "DISABLED",
-				'max404Humans_action' => "throttle",
-				'maxScanHits' => "DISABLED",
-				'maxScanHits_action' => "throttle",
-				'blockedTime' => "300"
-			)
-		),
-		array( //level 2
-			"checkboxes" => array(
-				"alertOn_critical" => true,
-				"alertOn_update" => false,
-				"alertOn_warnings" => true,
-				"alertOn_throttle" => false,
-				"alertOn_block" => true,
-				"alertOn_loginLockout" => true,
-				"alertOn_lostPasswdForm" => true,
-				"alertOn_adminLogin" => true,
-				"alertOn_nonAdminLogin" => false,
-				"liveTrafficEnabled" => true,
-				"advancedCommentScanning" => false,
-				"checkSpamIP" => false,
-				"spamvertizeCheck" => false,
-				"liveTraf_ignorePublishers" => true,
-				//"perfLoggingEnabled" => false,
-				"scheduledScansEnabled" => true,
-				"scansEnabled_public" => false,
-				"scansEnabled_heartbleed" => true,
-				"scansEnabled_core" => true,
-				"scansEnabled_themes" => false,
-				"scansEnabled_plugins" => false,
-				"scansEnabled_malware" => true,
-				"scansEnabled_fileContents" => true,
+				"scansEnabled_database" => true,
 				"scansEnabled_posts" => true,
 				"scansEnabled_comments" => true,
 				"scansEnabled_passwds" => true,
@@ -230,17 +157,20 @@ class wfConfig {
 				"addCacheComment" => false,
 				"disableCodeExecutionUploads" => false,
 				"allowHTTPSCaching" => false,
-				"debugOn" => false
+				"debugOn" => false,
+				'email_summary_enabled' => true,
+				'email_summary_dashboard_widget_enabled' => true,
+				'ssl_verify' => true,
 			),
 			"otherParams" => array(
-				'securityLevel' => '2',
+				'securityLevel' => '1',
 				"alertEmails" => "", "liveTraf_ignoreUsers" => "", "liveTraf_ignoreIPs" => "", "liveTraf_ignoreUA" => "",  "apiKey" => "", "maxMem" => '256', 'scan_exclude' => '', 'whitelisted' => '', 'bannedURLs' => '', 'maxExecutionTime' => '', 'howGetIPs' => '', 'actUpdateInterval' => '', 'alert_maxHourly' => 0, 'loginSec_userBlacklist' => '',
 				"neverBlockBG" => "neverBlockVerified",
-				"loginSec_countFailMins" => "240",
-				"loginSec_lockoutMins" => "240",
+				"loginSec_countFailMins" => "5",
+				"loginSec_lockoutMins" => "5",
 				'loginSec_strongPasswds' => 'pubs',
-				'loginSec_maxFailures' => "20",
-				'loginSec_maxForgotPasswd' => "20",
+				'loginSec_maxFailures' => "50",
+				'loginSec_maxForgotPasswd' => "50",
 				'maxGlobalRequests' => "DISABLED",
 				'maxGlobalRequests_action' => "throttle",
 				'maxRequestsCrawlers' => "DISABLED",
@@ -253,10 +183,13 @@ class wfConfig {
 				'max404Humans_action' => "throttle",
 				'maxScanHits' => "DISABLED",
 				'maxScanHits_action' => "throttle",
-				'blockedTime' => "300"
+				'blockedTime' => "300",
+				'email_summary_interval' => 'biweekly',
+				'email_summary_excluded_directories' => 'wp-content/cache,wp-content/wfcache,wp-content/plugins/wordfence/tmp',
+				'allowed404s' => "/favicon.ico\n/apple-touch-icon*.png\n/*@2x.png",
 			)
 		),
-		array( //level 3
+		array( //level 2
 			"checkboxes" => array(
 				"alertOn_critical" => true,
 				"alertOn_update" => false,
@@ -281,6 +214,7 @@ class wfConfig {
 				"scansEnabled_plugins" => false,
 				"scansEnabled_malware" => true,
 				"scansEnabled_fileContents" => true,
+				"scansEnabled_database" => true,
 				"scansEnabled_posts" => true,
 				"scansEnabled_comments" => true,
 				"scansEnabled_passwds" => true,
@@ -311,8 +245,102 @@ class wfConfig {
 				"startScansRemotely" => false,
 				"disableConfigCaching" => false,
 				"addCacheComment" => false,
+				"disableCodeExecutionUploads" => false,
 				"allowHTTPSCaching" => false,
-				"debugOn" => false
+				"debugOn" => false,
+				'email_summary_enabled' => true,
+				'email_summary_dashboard_widget_enabled' => true,
+				'ssl_verify' => true,
+			),
+			"otherParams" => array(
+				'securityLevel' => '2',
+				"alertEmails" => "", "liveTraf_ignoreUsers" => "", "liveTraf_ignoreIPs" => "", "liveTraf_ignoreUA" => "",  "apiKey" => "", "maxMem" => '256', 'scan_exclude' => '', 'whitelisted' => '', 'bannedURLs' => '', 'maxExecutionTime' => '', 'howGetIPs' => '', 'actUpdateInterval' => '', 'alert_maxHourly' => 0, 'loginSec_userBlacklist' => '',
+				"neverBlockBG" => "neverBlockVerified",
+				"loginSec_countFailMins" => "240",
+				"loginSec_lockoutMins" => "240",
+				'loginSec_strongPasswds' => 'pubs',
+				'loginSec_maxFailures' => "20",
+				'loginSec_maxForgotPasswd' => "20",
+				'maxGlobalRequests' => "DISABLED",
+				'maxGlobalRequests_action' => "throttle",
+				'maxRequestsCrawlers' => "DISABLED",
+				'maxRequestsCrawlers_action' => "throttle",
+				'maxRequestsHumans' => "DISABLED",
+				'maxRequestsHumans_action' => "throttle",
+				'max404Crawlers' => "DISABLED",
+				'max404Crawlers_action' => "throttle",
+				'max404Humans' => "DISABLED",
+				'max404Humans_action' => "throttle",
+				'maxScanHits' => "DISABLED",
+				'maxScanHits_action' => "throttle",
+				'blockedTime' => "300",
+				'email_summary_interval' => 'biweekly',
+				'email_summary_excluded_directories' => 'wp-content/cache,wp-content/wfcache,wp-content/plugins/wordfence/tmp',
+				'allowed404s' => "/favicon.ico\n/apple-touch-icon*.png\n/*@2x.png",
+			)
+		),
+		array( //level 3
+			"checkboxes" => array(
+				"alertOn_critical" => true,
+				"alertOn_update" => false,
+				"alertOn_warnings" => true,
+				"alertOn_throttle" => false,
+				"alertOn_block" => true,
+				"alertOn_loginLockout" => true,
+				"alertOn_lostPasswdForm" => true,
+				"alertOn_adminLogin" => true,
+				"alertOn_nonAdminLogin" => false,
+				"liveTrafficEnabled" => true,
+				"advancedCommentScanning" => false,
+				"checkSpamIP" => false,
+				"spamvertizeCheck" => false,
+				"liveTraf_ignorePublishers" => true,
+				//"perfLoggingEnabled" => false,
+				"scheduledScansEnabled" => true,
+				"scansEnabled_public" => false,
+				"scansEnabled_heartbleed" => true,
+				"scansEnabled_core" => true,
+				"scansEnabled_themes" => false,
+				"scansEnabled_plugins" => false,
+				"scansEnabled_malware" => true,
+				"scansEnabled_fileContents" => true,
+				"scansEnabled_database" => true,
+				"scansEnabled_posts" => true,
+				"scansEnabled_comments" => true,
+				"scansEnabled_passwds" => true,
+				"scansEnabled_diskSpace" => true,
+				"scansEnabled_options" => true,
+				"scansEnabled_dns" => true,
+				"scansEnabled_scanImages" => false,
+				"scansEnabled_highSense" => false,
+				"scansEnabled_oldVersions" => true,
+				"firewallEnabled" => true,
+				"blockFakeBots" => false,
+				"autoBlockScanners" => true,
+				"loginSecurityEnabled" => true,
+				"loginSec_lockInvalidUsers" => false,
+				"loginSec_maskLoginErrors" => true,
+				"loginSec_blockAdminReg" => true,
+				"loginSec_disableAuthorScan" => true,
+				"other_hideWPVersion" => true,
+				"other_noAnonMemberComments" => true,
+				"other_blockBadPOST" => false,
+				"other_scanComments" => true,
+				"other_pwStrengthOnUpdate" => true,
+				"other_WFNet" => true,
+				"other_scanOutside" => false,
+				"deleteTablesOnDeact" => false,
+				"autoUpdate" => false,
+				"disableCookies" => false,
+				"startScansRemotely" => false,
+				"disableConfigCaching" => false,
+				"addCacheComment" => false,
+				"disableCodeExecutionUploads" => false,
+				"allowHTTPSCaching" => false,
+				"debugOn" => false,
+				'email_summary_enabled' => true,
+				'email_summary_dashboard_widget_enabled' => true,
+				'ssl_verify' => true,
 			),
 			"otherParams" => array(
 				'securityLevel' => '3',
@@ -335,7 +363,10 @@ class wfConfig {
 				'max404Humans_action' => "throttle",
 				'maxScanHits' => "30",
 				'maxScanHits_action' => "throttle",
-				'blockedTime' => "1800"
+				'blockedTime' => "1800",
+				'email_summary_interval' => 'biweekly',
+				'email_summary_excluded_directories' => 'wp-content/cache,wp-content/wfcache,wp-content/plugins/wordfence/tmp',
+				'allowed404s' => "/favicon.ico\n/apple-touch-icon*.png\n/*@2x.png",
 			)
 		),
 		array( //level 4
@@ -363,6 +394,7 @@ class wfConfig {
 				"scansEnabled_plugins" => false,
 				"scansEnabled_malware" => true,
 				"scansEnabled_fileContents" => true,
+				"scansEnabled_database" => true,
 				"scansEnabled_posts" => true,
 				"scansEnabled_comments" => true,
 				"scansEnabled_passwds" => true,
@@ -393,8 +425,12 @@ class wfConfig {
 				"startScansRemotely" => false,
 				"disableConfigCaching" => false,
 				"addCacheComment" => false,
+				"disableCodeExecutionUploads" => false,
 				"allowHTTPSCaching" => false,
-				"debugOn" => false
+				"debugOn" => false,
+				'email_summary_enabled' => true,
+				'email_summary_dashboard_widget_enabled' => true,
+				'ssl_verify' => true,
 			),
 			"otherParams" => array(
 				'securityLevel' => '4',
@@ -417,7 +453,10 @@ class wfConfig {
 				'max404Humans_action' => "block",
 				'maxScanHits' => "10",
 				'maxScanHits_action' => "block",
-				'blockedTime' => "7200"
+				'blockedTime' => "7200",
+				'email_summary_interval' => 'biweekly',
+				'email_summary_excluded_directories' => 'wp-content/cache,wp-content/wfcache,wp-content/plugins/wordfence/tmp',
+				'allowed404s' => "/favicon.ico\n/apple-touch-icon*.png\n/*@2x.png",
 			)
 		)
 	);
@@ -438,6 +477,12 @@ class wfConfig {
 		}
 		if(self::get('other_scanOutside', false) === false){
 			self::set('other_scanOutside', 0);
+		}
+
+		if (self::get('email_summary_enabled')) {
+			wfActivityReport::scheduleCronJob();
+		} else {
+			wfActivityReport::disableCronJob();
 		}
 	}
 	public static function getExportableOptionsKeys(){
@@ -759,8 +804,11 @@ class wfConfig {
 		return true;
 	}
 	public static function enableAutoUpdate(){
-		wfConfig::set('autoUpdate', '1');	
-		wp_schedule_event(time(), 'daily', 'wordfence_daily_autoUpdate');
+		wfConfig::set('autoUpdate', '1');
+		wp_clear_scheduled_hook('wordfence_daily_autoUpdate');
+		if (is_main_site()) {
+			wp_schedule_event(time(), 'daily', 'wordfence_daily_autoUpdate');
+		}
 	}
 	public static function disableAutoUpdate(){
 		wfConfig::set('autoUpdate', '0');	
@@ -790,7 +838,9 @@ class wfConfig {
 				function show_message($msg = 'null'){}
 			}
 			*/
-			define('FS_METHOD', 'direct');
+			if(! defined('FS_METHOD')){ 
+				define('FS_METHOD', 'direct'); //May be defined already and might not be 'direct' so this could cause problems. But we were getting reports of a warning that this is already defined, so this check added. 
+			}
 			require_once(ABSPATH . 'wp-includes/update.php');
 			require_once(ABSPATH . 'wp-admin/includes/file.php');
 			wp_update_plugins();
@@ -825,11 +875,12 @@ Options -ExecCGI
 		$upload_dir = wp_upload_dir();
 		return $upload_dir['basedir'] . '/.htaccess';
 	}
-	
+
 	/**
 	 * Add/Merge .htaccess file in the uploads directory to prevent code execution.
 	 *
 	 * @return bool
+	 * @throws wfConfigException
 	 */
 	public static function disableCodeExecutionForUploads() {
 		$uploads_htaccess_file_path = self::_uploadsHtaccessFilePath();
@@ -848,27 +899,32 @@ Options -ExecCGI
 		}
 		return true;
 	}
-	
+
 	/**
 	 * Remove script execution protections for our the .htaccess file in the uploads directory.
 	 *
 	 * @return bool
+	 * @throws wfConfigException
 	 */
 	public static function removeCodeExecutionProtectionForUploads() {
 		$uploads_htaccess_file_path = self::_uploadsHtaccessFilePath();
 		if (file_exists($uploads_htaccess_file_path)) {
 			$htaccess_contents = file_get_contents($uploads_htaccess_file_path);
-			$htaccess_contents = str_replace(self::$_disable_scripts_htaccess, '', $htaccess_contents);
-			
-			$error_message = "Unable to remove code execution protections applied to the .htaccess file in the uploads directory.  Please check your permissions on that file.";
-			if (strlen(trim($htaccess_contents)) === 0) {
-				// empty file, remove it
-				if (!@unlink($uploads_htaccess_file_path)) {
+
+			// Check that it is in the file
+			if (strpos($htaccess_contents, self::$_disable_scripts_htaccess) !== false) {
+				$htaccess_contents = str_replace(self::$_disable_scripts_htaccess, '', $htaccess_contents);
+
+				$error_message = "Unable to remove code execution protections applied to the .htaccess file in the uploads directory.  Please check your permissions on that file.";
+				if (strlen(trim($htaccess_contents)) === 0) {
+					// empty file, remove it
+					if (!@unlink($uploads_htaccess_file_path)) {
+						throw new wfConfigException($error_message);
+					}
+
+				} elseif (@file_put_contents($uploads_htaccess_file_path, $htaccess_contents, LOCK_EX) === false) {
 					throw new wfConfigException($error_message);
 				}
-				
-			} elseif (@file_put_contents($uploads_htaccess_file_path, $htaccess_contents, LOCK_EX) === false) {
-				throw new wfConfigException($error_message);
 			}
 		}
 		return true;
